Add `persistence.Provider` and move providers into subfolder.'
