--- persistence/boltdb/doc.go
@@ -1,2 +1,0 @@
-// Package boltdb is a BoltDB (bbolt) persistence driver.
-package boltdb
