--- persistence/memory/doc.go
@@ -1,2 +1,0 @@
-// Package memory is an in-memory persistence driver.
-package memory
