+++ persistence/provider/boltdb/doc.go
@@ -1,0 +1,2 @@
+// Package boltdb is a BoltDB (bbolt) persistence provider.
+package boltdb
