+++ persistence/provider/memory/doc.go
@@ -1,0 +1,2 @@
+// Package memory is an in-memory persistence provider.
+package memory
