+++ persistence/provider/memory/stream.go
