+++ persistence/provider/sql/doc.go
@@ -1,0 +1,3 @@
+// Package sql is an SQL-based persistence providers with drivers for several
+// popular SQL database systems.
+package sql
