+++ persistence/provider/sql/driver/mysql/doc.go
@@ -1,0 +1,2 @@
+// Package mysql is a MySQL driver for the SQL persistence provider.
+package mysql
