+++ persistence/provider/sql/driver/mysql/stream_test.go
@@ -1,17 +1,16 @@
-package sqlite_test
+package mysql_test
 
 import (
 	"context"
 	"database/sql"
-	"strings"
 	"time"
 
 	"github.com/dogmatiq/infix/envelope"
 	"github.com/dogmatiq/infix/internal/sqltest"
 	"github.com/dogmatiq/infix/internal/x/sqlx"
 	"github.com/dogmatiq/infix/persistence/internal/streamtest"
-	infixsql "github.com/dogmatiq/infix/persistence/sql"
-	. "github.com/dogmatiq/infix/persistence/sql/driver/sqlite"
+	infixsql "github.com/dogmatiq/infix/persistence/provider/sql"
+	. "github.com/dogmatiq/infix/persistence/provider/sql/driver/mysql"
 	"github.com/dogmatiq/linger/backoff"
 	"github.com/dogmatiq/marshalkit"
 	. "github.com/onsi/ginkgo"
@@ -23,12 +22,9 @@ var _ = Describe("type StreamDriver (standard test suite)", func() {
 
 	streamtest.Declare(
 		func(ctx context.Context, m marshalkit.Marshaler) streamtest.Config {
-			db = sqltest.Open("sqlite3")
+			db = sqltest.Open("mysql")
 
 			err := DropSchema(ctx, db)
-			if err != nil && strings.Contains(err.Error(), "CGO_ENABLED") {
-				Skip(err.Error())
-			}
 			Expect(err).ShouldNot(HaveOccurred())
 
 			err = CreateSchema(ctx, db)
