+++ persistence/provider/sql/driver/postgres/doc.go
@@ -1,0 +1,2 @@
+// Package postgres is a PostgreSQL driver for the SQL persistence provider.
+package postgres
