+++ persistence/provider/sql/driver/postgres/stream_test.go
@@ -1,6 +1,4 @@
-// +build cgo
-
-package sql_test
+package postgres_test
 
 import (
 	"context"
@@ -11,31 +9,31 @@ import (
 	"github.com/dogmatiq/infix/internal/sqltest"
 	"github.com/dogmatiq/infix/internal/x/sqlx"
 	"github.com/dogmatiq/infix/persistence/internal/streamtest"
-	. "github.com/dogmatiq/infix/persistence/sql"
-	"github.com/dogmatiq/infix/persistence/sql/driver/sqlite"
+	infixsql "github.com/dogmatiq/infix/persistence/provider/sql"
+	. "github.com/dogmatiq/infix/persistence/provider/sql/driver/postgres"
 	"github.com/dogmatiq/linger/backoff"
 	"github.com/dogmatiq/marshalkit"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 )
 
-var _ = Describe("type Stream (standard test suite)", func() {
+var _ = Describe("type StreamDriver (standard test suite)", func() {
 	var db *sql.DB
 
 	streamtest.Declare(
 		func(ctx context.Context, m marshalkit.Marshaler) streamtest.Config {
-			db = sqltest.Open("sqlite3")
+			db = sqltest.Open("postgres")
 
-			err := sqlite.DropSchema(ctx, db)
+			err := DropSchema(ctx, db)
 			Expect(err).ShouldNot(HaveOccurred())
 
-			err = sqlite.CreateSchema(ctx, db)
+			err = CreateSchema(ctx, db)
 			Expect(err).ShouldNot(HaveOccurred())
 
-			stream := &Stream{
+			stream := &infixsql.Stream{
 				ApplicationKey:  "<app-key>",
 				DB:              db,
-				Driver:          sqlite.StreamDriver{},
+				Driver:          StreamDriver{},
 				Marshaler:       m,
 				BackoffStrategy: backoff.Constant(10 * time.Millisecond),
 			}
