+++ persistence/provider/sql/driver/sqlite/doc.go
@@ -1,0 +1,2 @@
+// Package sqlite is a SQlite v3 for the SQL persistence provider.
+package sqlite
