--- persistence/sql/doc.go
@@ -1,2 +1,0 @@
-// Package sql contains various SQL-based persistence drivers.
-package sql
