--- persistence/sql/driver/mysql/doc.go
@@ -1,2 +1,0 @@
-// Package mysql is a MySQL (and compatible) persistence driver.
-package mysql
