--- persistence/sql/driver/postgres/doc.go
@@ -1,2 +1,0 @@
-// Package postgres is a PostgreSQL persistence driver.
-package postgres
