--- persistence/sql/driver/sqlite/doc.go
@@ -1,2 +1,0 @@
-// Package sqlite is a SQLite (version 3) persistence driver.
-package sqlite
